# perf(context): optimize getMapFromFormData performance