@ddevsr 2025-09-01
I can confirm

After:
```bash
goos: windows
goarch: amd64
pkg: github.com/PHPDevsr/gin-forked
cpu: Intel(R) Core(TM) i5-3570 CPU @ 3.40GHz
BenchmarkGetMapFromFormData/Small_Bracket-4         	 3426664	       338.7 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Small_Names-4           	 3546652	       326.5 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Bracket-4        	 2140892	       551.5 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Names-4          	 2342815	       510.4 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Other-4          	 2525497	       462.6 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Large_Bracket-4         	   94688	     12933 ns/op	    9368 B/op	      11 allocs/op
BenchmarkGetMapFromFormData/Large_Names-4           	  160042	      7721 ns/op	    4472 B/op	       9 allocs/op
BenchmarkGetMapFromFormData/Large_Other-4           	  244315	      5005 ns/op	    2136 B/op	       7 allocs/op
BenchmarkGetMapFromFormData/WorstCase_Bracket-4     	  601768	      1978 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/ShortKeys_Bracket-4     	 3754584	       321.5 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Empty_Key-4             	 9444721	       127.4 ns/op	      48 B/op	       1 allocs/op
PASS
ok  	github.com/PHPDevsr/gin-forked	13.299s
```

Before:
```bash
goos: windows
goarch: amd64
pkg: github.com/PHPDevsr/gin-forked
cpu: Intel(R) Core(TM) i5-3570 CPU @ 3.40GHz
BenchmarkGetMapFromFormData/Small_Bracket-4         	 3494000	       343.1 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Small_Names-4           	 3361057	       352.5 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Bracket-4        	 2192718	       549.2 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Names-4          	 2171438	       554.4 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Medium_Other-4          	 2320060	       545.1 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Large_Bracket-4         	   84973	     13897 ns/op	    9368 B/op	      11 allocs/op
BenchmarkGetMapFromFormData/Large_Names-4           	  140596	      8698 ns/op	    4472 B/op	       9 allocs/op
BenchmarkGetMapFromFormData/Large_Other-4           	  203253	      6168 ns/op	    2136 B/op	       7 allocs/op
BenchmarkGetMapFromFormData/WorstCase_Bracket-4     	  556111	      2049 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/ShortKeys_Bracket-4     	 3435433	       344.4 ns/op	     336 B/op	       2 allocs/op
BenchmarkGetMapFromFormData/Empty_Key-4             	 7792040	       153.5 ns/op	      48 B/op	       1 allocs/op
PASS
ok  	github.com/PHPDevsr/gin-forked	13.384s
```